+++ src/App.js
@@ -2,11 +2,7 @@ import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import RootLayout from "./router/rootLayout";
 import MainPage from "./pages/mainPage";
 import ListPage from "./pages/listPage";
-import GlobalStyle from "./assets/style/GlobalStyle";
-import { ChakraProvider } from "@chakra-ui/react";
-import theme from "./assets/style/theme";
-import RankingPage from "./pages/rankingPage";
-import MyPage from "./pages/myPage";
+import GlobalStyle from "./assets/image/style/GlobalStyle";
 
 const router = createBrowserRouter([
   {
@@ -21,24 +17,16 @@ const router = createBrowserRouter([
         path: "list",
         element: <ListPage />,
       },
-      {
-        path: "ranking",
-        element: <RankingPage />,
-      },
-      {
-        path: "mypage",
-        element: <MyPage />,
-      },
     ],
   },
 ]);
 
 function App() {
   return (
-    <ChakraProvider theme={theme}>
+    <>
       <GlobalStyle />
       <RouterProvider router={router} />
-    </ChakraProvider>
+    </>
   );
 }
