--- src/components/input.tsx
@@ -1,20 +1,0 @@
-import { Search2Icon } from "@chakra-ui/icons";
-import {
-  Input as ChakraInput,
-  InputGroup,
-  InputRightElement,
-} from "@chakra-ui/react";
-
-export function SearchInput() {
-  return (
-    <InputGroup>
-      <ChakraInput
-        placeholder="상품명을 입력해주세요"
-        focusBorderColor="pink.100"
-      />
-      <InputRightElement>
-        <Search2Icon color="pink.100" />
-      </InputRightElement>
-    </InputGroup>
-  );
-}
