+++ src/router/rootLayout.tsx
@@ -1,10 +1,8 @@
-import Header from "../components/header";
 import { Outlet } from "react-router-dom";
 
 export default function RootLayout() {
   return (
     <>
-      <Header />
       <main>
         <Outlet />
       </main>
